requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-whois==0.8.0
urllib3==2.0.7
colorama==0.4.6
dnspython==2.4.2
pyopenssl==24.0.0
tld==0.13.0
python-nmap==0.7.1
validators==0.22.0
